requirements.txt
Celery
Django==4.1.6
djangorestframework
django-debug-toolbar
django-extensions==3.2.1
ipython
django-admin-rangefilter
django-import-export
django-crispy-forms==2.0
requests
django-filter
djangorestframework-simplejwt
drf-yasg
pymemcache
gunicorn

#flake8
flake8
flake8-builtins
flake8-print

#pytest
pytest
pytest-django
pytest-cov
pytest-mock
